requirements.txt
fastapi
uvicorn
streamlit
faiss-cpu
sentence-transformers
requests
python-multipart
pypdf
numpy